fiing test failure
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@12110 f5854215-dd47-0410-b2c4-cdd35faa7885

+++ motion_planning/planning_models/src/load_mesh.cpp
@@ -185,7 +185,7 @@ namespace planning_models
 	    pos += 4;
 	    
 	    // make sure we have read enough data
-	    if (50 * numTriangles + 84 <= fileSize)
+	    if ((long)(50 * numTriangles + 84) <= fileSize)
 	    {
 		std::vector<btVector3> vertices;
